requirements.txt
APScheduler==3.6.3
astroid==2.11.4
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
clarifai-grpc==8.0.0
dill==0.3.4
emoji==1.7.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
protobuf==3.20.1
pylint==2.13.8
PySocks==1.7.1
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wrapt==1.14.1